Split VerifyFlow from index
Avoid dependency cycle, make room for more index-exported
