+++ app/javascript/packages/verify-flow/steps/personal-key-confirm/index.ts
@@ -1,6 +1,6 @@
 import { t } from '@18f/identity-i18n';
 import type { FormStep } from '@18f/identity-form-steps';
-import type { VerifyFlowValues } from '../..';
+import type { VerifyFlowValues } from '../../verify-flow';
 import form from './personal-key-confirm-step';
 
 export default {
